requirements.txt
awscli==1.16.92
boto3==1.9.74
botocore==1.12.82
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==1.11.18
django-cors-headers==2.4.0
djangorestframework==3.9.0
docutils==0.14
futures==3.2.0
gunicorn==19.9.0
idna==2.8
jmespath==0.9.3
psycopg2==2.7.6.1
pyasn1==0.4.5
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
